requirements.txt
fastapi
uvicorn
groq
python-dotenv
pyautogui
screen-brightness-control
